feat: home page to search and list pessoas

+++ lib/app/data/models/atributos_fisicos_model.dart
@@ -1,6 +1,6 @@
 class AtributosFisicos {
   double? altura;
-  int? peso;
+  double? peso;
   String? sexo;
   String? tipoSanguineo;
